backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
DateTime==5.5
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.16.0
idna==3.10
jwt==1.3.1
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
rsa==4.9.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
zope.interface==7.2
